renaming stochastic controller to LocobotFPSAgentController; preventing agentMode=default + agentControllerType=stochastic + unit test
