+++ unity/Assets/Scripts/LocobotFPSAgentController.cs.meta
@@ -1,3 +1,11 @@
-﻿fileFormatVersion: 2
-guid: 62944b1c50894c5c9ee6581cf645ebcb
-timeCreated: 1630379176
+fileFormatVersion: 2
+guid: 6839fe93e39084f7ba1b70ae71c8d153
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
