--- unity/Assets/Scripts/StochasticRemoteFPSAgentController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6839fe93e39084f7ba1b70ae71c8d153
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
